+++ seeds/molecular/__init__.py
@@ -8,6 +8,7 @@ from .m01_hierarchy_builder import HierarchyBuilder, create_hierarchy_builder
 from .m02_causality_detector import CausalityDetector, create_causality_detector
 from .m03_pattern_completer import PatternCompleter
 from .m04_spatial_transformer import SpatialTransformer, create_spatial_transformer
+from .m06_context_integrator import ContextIntegrator, create_context_integrator
 
 __all__ = [
     "HierarchyBuilder",
@@ -17,5 +18,7 @@ __all__ = [
     "PatternCompleter",
     "SpatialTransformer",
     "create_spatial_transformer",
+    "ContextIntegrator",
+    "create_context_integrator",
 ]